Tidy up the scaffolded code
